Separate out net request implementations
